LayoutTests/media/modern-media-controls/macos-fullscreen-media-controls/macos-fullscreen-media-controls-buttons-styles-expected.txt

Testing the MacOSFullscreenMediaControls computed styles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



RewindButton
PASS getComputedStyle(button.element).webkitMaskPositionY became "17px"
PASS getComputedStyle(button.element).left is "27px"
PASS getComputedStyle(button.element).height is "44px"

PlayPauseButton
PASS getComputedStyle(button.element).webkitMaskPositionY became "12px"
PASS getComputedStyle(button.element).left is "78px"
PASS getComputedStyle(button.element).height is "44px"

ForwardButton
PASS getComputedStyle(button.element).webkitMaskPositionY became "17px"
PASS getComputedStyle(button.element).left is "127px"
PASS getComputedStyle(button.element).height is "44px"

AirplayButton
PASS getComputedStyle(button.element).webkitMaskPositionY became "18px"
PASS getComputedStyle(button.element).left is "12px"
PASS getComputedStyle(button.element).height is "44px"

PiPButton
PASS getComputedStyle(button.element).webkitMaskPositionY became "18px"
PASS getComputedStyle(button.element).left is "44px"
PASS getComputedStyle(button.element).height is "44px"

TracksButton
PASS getComputedStyle(button.element).webkitMaskPositionY became "18px"
PASS getComputedStyle(button.element).left is "79px"
PASS getComputedStyle(button.element).height is "44px"

FullscreenButton
PASS getComputedStyle(button.element).webkitMaskPositionY became "18px"
PASS getComputedStyle(button.element).left is "111px"
PASS getComputedStyle(button.element).height is "44px"
PASS successfullyParsed is true

TEST COMPLETE
